Fix 유저 수정 삭제 본인만 가능토록 수정

+++ src/main/java/com/iso/logus/domain/user/controller/UserController.java
@@ -2,6 +2,8 @@ package com.iso.logus.domain.user.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,6 +17,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.iso.logus.domain.user.dto.UserDto;
 import com.iso.logus.domain.user.service.UserService;
+import com.iso.logus.global.exception.AccessDeniedException;
+import com.iso.logus.global.jwt.JwtTokenProvider;
 
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
@@ -26,6 +30,7 @@ import lombok.RequiredArgsConstructor;
 public class UserController {
 	
 	private final UserService userService;
+	private final JwtTokenProvider jwtTokenProvider;
 	
 	@GetMapping(value = "/{uid}")
 	@ApiImplicitParams({
@@ -45,7 +50,9 @@ public class UserController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 access_token", required = false, dataType = "String", paramType = "header")
 	})
-	public void changeUserName(@PathVariable String uid, @RequestBody UserDto.ChangeNameRequest changeNameRequest) {
+	public void changeUserName(@PathVariable String uid, @RequestBody UserDto.ChangeNameRequest changeNameRequest, HttpServletRequest request) {
+		if(!jwtTokenProvider.validateUser(uid, request))
+			throw new AccessDeniedException();
 		userService.changeUserName(uid, changeNameRequest);
 	}
 	
@@ -53,7 +60,9 @@ public class UserController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 access_token", required = false, dataType = "String", paramType = "header")
 	})
-	public void deleteUserByUid(@PathVariable String uid) {
+	public void deleteUserByUid(@PathVariable String uid, HttpServletRequest request) {		
+		if(!jwtTokenProvider.validateUser(uid, request))
+			throw new AccessDeniedException();
 		userService.deleteUserByUid(uid);
 	}
